pyproject.toml
[project]
name = "detec-facial"
version = "1.0.0"
description = "Sistema multimodal de triagem para saúde da mulher (vídeo + áudio)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "numpy>=1.24.0",
    "librosa>=0.10.0",
    "ultralytics>=8.0.0",
    "openai-whisper>=20230918",
]
